+++ src/cachefile.hh
@@ -1,4 +1,4 @@
-/* $Id: cachefile.hh,v 1.3 2005-07-02 22:05:03 atterer Exp $ -*- C++ -*-
+/* $Id: cachefile.hh,v 1.4 2005-07-21 11:31:43 atterer Exp $ -*- C++ -*-
   __   _
   |_) /|  Copyright (C) 2001-2002  |  richard@
   | \/¯|  Richard Atterer          |  atterer.net
@@ -94,7 +94,7 @@ public:
       only valid until the next database operation. */
   Status findName(const byte*& resultData, size_t& resultSize,
                   const string& fileName,
-                  long long int& resultFileSize, time_t& resultMtime);
+                  off_t& resultFileSize, time_t& resultMtime);
 
   /** Insert/overwrite entry for the given file (name must be
       absolute, file must have the supplied mtime and size). The data
